Desktop/job-scraper/server/requirements.txt
selenium
undetected-chromedriver
webdriver-manager
requests
beautifulsoup4
python-dotenv
fake-useragent
fastapi
uvicorn
python-dotenv 
psycopg2